Update to include ranges and countries

+++ src/WebDemo/Models/AlertRange.cs
@@ -10,8 +10,8 @@ namespace WebDemo.Models
     public class AlertRange
     {
         public int Alert_Type_Id { get; set; }
-        public int Capture_PC_Id { get; set; }
-        public int Recording_Id { get; set; }
+        public long Capture_PC_Id { get; set; }
+        public long Recording_Id { get; set; }
         [JsonConverter(typeof(TimeFrame))]
         public TimeSpan Time_Spent_In_Alert_State { get; set; }
         public bool Currently_In_Alert_State { get; set; }
